Install react router dependencies
