+++ Module_six/hello-react-front-end/src/App.js
@@ -1,14 +1,15 @@
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { Provider } from "react-redux";
-import store from "./redux/configureStore";
-import Greeting from "./Greeting";
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from './redux/configureStore';
+import Greeting from './Greeting';
+
 function App() {
   return (
     <Provider store={store}>
       <BrowserRouter>
         <Routes>
-        <Route path="/" element={<Greeting />} />
+          <Route path="/" element={<Greeting />} />
         </Routes>
       </BrowserRouter>
     </Provider>
