+++ Module_six/hello-react-front-end/src/redux/configureStore.js
@@ -1,13 +1,13 @@
-import { combineReducer } from "redux";
-import { configureStore } from "@reduxjs/toolkit";
-import helloReducer from "./greet";
+import { combineReducer } from 'redux';
+import { configureStore } from '@reduxjs/toolkit';
+import helloReducer from './greet';
 
-const reducer = combineReducer ({
-    helloReducer,
-})
+const reducer = combineReducer({
+  helloReducer,
+});
 
-const store = configureStore ({
-    reducer,
-})
+const store = configureStore({
+  reducer,
+});
 
 export default store;
